backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.36
aiosqlite>=0.20.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx[http2]>=0.28.0
beautifulsoup4>=4.12.0
lxml>=5.3.0
cssutils>=2.11.0
bleach>=6.2.0
python-multipart>=0.0.18
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
websockets>=14.1
